pyproject.toml
[project]
name = "timeseries-pro"
version = "2.0.0"
description = "Enterprise-grade time series forecasting platform with advanced data science capabilities"
requires-python = ">=3.11"
dependencies = [
    # Core framework (existing)
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.41",
    "statsmodels>=0.14.4",
    "werkzeug>=3.1.3",
    
    # ML Libraries (existing)
    "lightgbm>=4.6.0",
    "prophet>=1.1.6",
    "psycopg2-binary>=2.9.10",
    "xgboost>=2.0.0",
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "pytorch-forecasting>=1.0.0",
    
    # External data integration
    "holidays>=0.74",
    "requests>=2.31.0",
    
    # Advanced analytics and optimization
    "optuna>=3.4.0",
    "scikit-optimize>=0.9.0",
    "featuretools>=1.26.0",
    "tsfresh>=0.20.1",
    
    # Async processing and caching
    "celery>=5.3.4",
    "redis>=5.0.1",
    "kombu>=5.3.4",
    
    # WebSocket support
    "flask-socketio>=5.3.6",
    "python-socketio>=5.10.0",
    
    # Monitoring and observability
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    
    # Security and utilities
    "cryptography>=41.0.7",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.2",
    
    # Data processing enhancements
    "polars>=0.20.2",
    "pyarrow>=14.0.2",
    
    # Additional ML utilities
    "imbalanced-learn>=0.11.0",
    "category-encoders>=2.6.3",
    
    # Data profiling and reporting
    "sweetviz>=2.3.0",
]
